Refactor file scanning
